(feat) make unit_value to float

+++ database/migrations/2020_11_01_110153_create_products_table.php
@@ -31,7 +31,7 @@ class CreateProductsTable extends Migration
             $table->id();
             $table->foreignId('user_id')->constrained('users');
             $table->foreignId('product_id')->constrained('products');
-            $table->integer('unit_value');
+            $table->float('unit_value');
             $table->timestamps();
         });
     }
